Revert "Change group information"
