+++ front-end/src/App.js
@@ -14,8 +14,7 @@ import Go from './component/Go';
 function App() {
   return (
     <div className='App p-6 max-h-screen flex justify-center min-w-screen max-w-screen md: max-w-800'>
-      {/* <div className='min-w-full max-w-full'> */}
-      <div>
+      <div className='min-w-full max-w-full'>
         <div className='routes h-2/3'>
           <Routes>
             <Route path='/' element={<CreateNewTeam />} />
