+++ front-end/src/component/Bubble.js
