+++ front-end/src/component/Button.js
@@ -4,7 +4,6 @@ import {useNavigate} from "react-router-dom";
 import Home from './Home';
 import Popup from 'reactjs-popup';
 import Space from './Space';
-import Memcpy from './Memcpy';
 
 function B0({str})
 {
@@ -50,7 +49,7 @@ function B4({str})
     let navigate = useNavigate();
     return(
         <div>
-            <button className = 'b4' onClick = {() => str.left ? navigate('/') : navigate('/Group')} style = {{left: str.left ? '-80px' : '80px', top: str.left ? '' : '-26px'}}>{str.str}</button>
+            <button className = 'b4' onClick = {() => str.left ? navigate('/') : navigate('/team_info')} style = {{left: str.left ? '-80px' : '80px', top: str.left ? '' : '-26px'}}>{str.str}</button>
         </div>
     );
 }
@@ -104,26 +103,7 @@ function B9({str})
     );
 }
 
-function B10({str})
-{
-    let input = 'input0';
-
-    function Input_function()
-    {
-        return(
-            <input id = {input} className = 'b3' type = 'text' value = {str.placeholder} style = {{marginLeft: '-2px', marginTop: '-2px'}}></input>
-        );
-    }
-
-    return(
-        <div>
-            <label className = 'b3 block mb-2' style = {{textAlign: str.text_align}}>{str.str}</label>
-            <button className = 'b7' onClick = {() => Memcpy(input)}><Input_function></Input_function></button>
-        </div>
-    );
-}
-
-let component_array = [B0, B1, B2, B3, B4, B5, B6, B7, B8, B9, B10];
+let component_array = [B0, B1, B2, B3, B4, B5, B6, B7, B8, B9];
 
 function Button({str_array, type})
 {
