+++ front-end/src/component/Canvas.js
