+++ front-end/src/component/CreateNewTeam.css
@@ -5,9 +5,3 @@ img.img0
     margin-left: auto;
     margin-right: auto;
 };
-
-.link0
-{
-    text-decoration: none;
-    color: #723d46;
-}
