+++ front-end/src/component/CreateNewTeam.js
@@ -4,8 +4,6 @@ import mainback from '../image/mainback.png';
 import bowl from '../image/bowl.png';
 import '../App.css';
 import './CreateNewTeam.css';
-import {Link} from 'react-router-dom';
-import Space from './Space';
 
 function CreateNewTeam()
 {
@@ -14,12 +12,10 @@ function CreateNewTeam()
         <div className = 'App_body' style = {{backgroundImage: `url(${mainback})`}}>
             <form>
                 <img src = {bowl} alt = 'bowl' className = 'img0 align-center'></img><br></br>
-                <Button str_array = {[{str: 'Create a Group', link: 'Home'}]} type = {7}></Button><br></br>
+                <Button str_array = {[{str: 'Create a Group', link: 'Home'}]} type = {7}></Button>
                 <Button str_array = {['Join a Group']} type = {1}></Button>
-                <Button str_array = {['Enter']} type = {2}></Button><br></br><br></br><br></br>
-                <Link to = 'Signup' style = {{textDecoration: 'none', color: '#723d46', fontSize: '20px'}}>Sign Up</Link>
-                <Space n = {20}></Space>
-                <Link to = 'Login' style = {{textDecoration: 'none', color: '#723d46', fontSize: '20px'}}>Log In</Link>
+                <Button str_array = {['Enter']} type = {2}></Button><br></br><br></br>
+                <Button str_array = {[{str: 'Be a Long-Term User', link: 'Signup'}, {str: 'Long-Term User Log In', link: 'Login'}]} type = {7}></Button>
             </form>
         </div>
     </div>
