--- front-end/src/component/Memcpy.jsx
@@ -1,9 +1,0 @@
-function Memcpy(input)
-{
-    let copy = document.getElementById(input);
-    copy.select();
-    navigator.clipboard.writeText(copy.value);
-    alert('Link copied: ' + copy.value);
-}
-
-export default Memcpy;
