+++ front-end/src/component/Space.js
