+++ front-end/src/component/Vote.js
